+++ CHANGELOG.md
@@ -8,7 +8,8 @@ Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 ## [Unreleased]
 #### Added
-- `Document` class
+- `Document` and `upload-documents`
+- `VerificationManager` helper
 
 ## [1.13.1-rc.0] - 2020-06-02
 ### Changed